% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_sum_exp}
\alias{log_sum_exp}
\title{log_sum_exp}
\usage{
log_sum_exp(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
The natural logarithm of sum of \code{exp(x)}
}
\description{
A numerically stable version of \code{log(sum(exp(x)))}
}
