\name{CAPPMx-package}
\alias{CAPPMx-package}
\alias{CAPPMx}
\docType{package}
\title{
\packageTitle{CAPPMx}
}
\description{
\packageDescription{CAPPMx}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CAPPMx}
\packageIndices{CAPPMx}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{CAPPMx}

Maintainer: \packageMaintainer{CAPPMx}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
