\name{CAPPMx-package}
\alias{CAPPMx-package}
\alias{CAPPMx}
\docType{package}
\title{
\packageTitle{CAPPMx}
}
\description{
An implementation of the Bayesian Nonparametric Common Atoms Regression by \insertRef{chandra_GBM22;textual}{CAPPMx}.
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CAPPMx}
\packageIndices{CAPPMx}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{CAPPMx}

Maintainer: \packageMaintainer{CAPPMx}
}
\references{
\insertRef{chandra_GBM22}{CAPPMx}
}
\keyword{ CAPPMx }
\examples{
## See the package-vignette file.
}